Update to latest Spigot
The BaseChunkGenerator.getBiomesForChunk method is now deprecated, use getBlocksForChunk().getBiome instead. This was necessary because Bukkit deprecated the BiomeGrid.

Now that Bukkit has an actually good world generator api, most of WorldGeneratorApi is no longer necessary. :smiley: I'll have a look at what needs to remain, and what parts of the API can be removed in the future.